Dynamic search of ek keys with ekCertificates (#468)
* Dynamic search of ek keys with ekCertificates

This patch allows ek keys to be searched in additional
persistent key handle locations. And to return the actual
Certificate and handle mapping when the keys in the handles
do not match the default values.

Unifies creation of EK and allows listing of high certs
if present.

Rationale:
During the deployment of some newer lenovo systems the default
rsa handle comes with a 3k key which prevents attestation
from working as built. Thus the need for the library to create the
2k key in a separate location in case the default location is
already occupied and there is no 2k on the standard locations.

* addressing comments p1

* Adding test using local tpm.

* Added documentation in the choice of keyhandles

* Remove p384 and ek 3072 as not relevant to this change

* Added test of EkCertificates with simulator

Mostly used internal interfaces as simulator does
not have an EkCert in nvram location.

* Cleaned up comments and added more docs

* fixing nits from liamjm's review